Add preliminary support for Windows. - Bump version to 1.1.1 - TODO: Windows version has hardcoded path to plantuml.jar (c:\bin)
Issue on Windows is that it is essentially impossible to wrap the execution of PlantUML.jar in a batch file as these do not really function properly in a pipe. The optimum approach would be to search the path (and class path?) for PlantUML.jar, as this would eliminate the need for a shell script on Unix as well.